tweaked sickle models
turned the blessed model into a child model to reduce file size. changed the model and colors to better fit the shape and look.
side note: figured out why models are half hidden in the GE, in the display translation for GUI, if any part of the model goes into the -Z value it clips behind the GEs interface.